+++ src/main/java/sg4e/ff4stats/Battle.java
@@ -23,6 +23,7 @@ import java.util.Objects;
 import org.apache.commons.text.WordUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import sg4e.ff4stats.csv.*;
 
 /**
  *
