+++ src/main/java/sg4e/ff4stats/csv/CSVParser.java
@@ -14,7 +14,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package sg4e.ff4stats;
+package sg4e.ff4stats.csv;
 
 import java.io.IOException;
 import java.io.InputStream;
