+++ src/main/java/sg4e/ff4stats/fe/FlagVersion.java
@@ -26,8 +26,7 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import sg4e.ff4stats.RecordParser;
-import sg4e.ff4stats.CSVParser;
+import sg4e.ff4stats.csv.*;
 
 /**
  *
