+++ src/main/java/sg4e/ff4stats/party/LevelData.java
@@ -31,8 +31,7 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import sg4e.ff4stats.RecordParser;
-import sg4e.ff4stats.CSVParser;
+import sg4e.ff4stats.csv.*;
 
 /**
  *
