+++ src/main/java/sg4e/ff4stats/party/Weapon.java
@@ -26,9 +26,7 @@ import java.util.List;
 import java.util.Set;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import sg4e.ff4stats.RecordParser;
-import sg4e.ff4stats.CSVParser;
-import sg4e.ff4stats.RecordParser;
+import sg4e.ff4stats.csv.*;
 
 /**
  *
